Add support of getting keys from config
Instead of predefined the keys in the python script,
allow the keys to be parsed from the config file. So
now, we can get the config file from our dashboard and
set the application key and network key accordingly.

For the time being, we mainly use the application key
only.

Signed-off-by: Chun-Yeow Yeoh <yeohchunyeow@gmail.com>